refactor: relocate mocks
